NCBC-764: Implement DNS-SRV
MOTIVATION
----------
DNS SRV records are used to manage changing application DNS endpoints
without having to change application config. Instead of storing DNS
names or IP addresses in config, a DNS SRV record is created and when
queried, describes a list of DNS registrations for use with a particular
service.

The .NET SDK does not currnetly support bootstrapping using an external
service, such as DNS SRV, and it should to provide parity with the other
SDKs.

NOTE: The .NET framework does not provide a consistent method to query
DNS SRV records. The typical way to do this was to use a third party
tool that likely imported a windows binary to allow the query to
execute. This is not ideal and especially now .NET Core targets
multiple OS platforms. Because of this, this implementation leaves the
majority of the work of retriveing the URIs to the consuming application.
This change provides the framework for allowing the bootstrapping to
happen and this has had the side effect that the bootstrapping process is
not limited to DNS SRV, but any external service could provide the
bootstrap URIs.

MODIFICATIONS
-------------
Add IServerResolver interface for consuming applications to implement to
return valid URIs.
Add ServerResolverElement to hold the ServerResolver configuration.
Add ServerResolverUtil which is used during the Bootstrap process to check
the Couchbase config for a customised ServerResolver Type to bootstrap with.
Updated CouchbaseClientDefinition and CouchbaseClientSection with
appropriate properties.
Add DefaultServerResolver as a basic illustration for how to implement
an IServerResolver which returns a single localhost URI.

Add Unit Tests to cover the following scenarios:
- ServerResolver is not present
- ServerResolver is not valid (doesn't implement inteface / class doesn't
exist)
- ServerResolver does not return any URIs
- ServerResolver returns URIs
- ServerResolver is loaded from config file

RESULT
------
A custom implementation to source a seed list of server URIS to bootstrap
with can be provided as part of the client configuration. One implementation
of this could be DNS SRV.

Change-Id: I76a0a4f95fb5adb2184387cfdb6b97b823dcf1bf
Reviewed-on: http://review.couchbase.org/69203
Tested-by: buildbot <build@couchbase.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>