NCBC-1204: CouchbaseClientDefinition should use interface for Buckets property
MOTIVATION
----------
The .NET Couchbase supports configuration to be provided through a JSON
file at start-up. This is provided using IConfigurationProvider from
Microsoft.Extensions.Configuration. However, this implementation does
not support interface types when binding properties.

The original intention of this ticket was to replace the Buckets
property of CouchbaseClientDefinition to use List<IBucketConfiguration>.
However now the above has come to light, we feel it best to not to
modify the property type or we would lose the feature of loading config
from a JSON file.

NODIFICATIOS
------------
Add a unit test to prove that the current implementation can correctly
bind the Buckets property of the config and to act as a safeguard
against changing it in the future.

RESULT
------
Unit test coverage and explaination of why we chose not to change the
Buckets propety type have been added to the source code.

Change-Id: I477f69875799122b48362c6ad943de0ffc9991f9
Reviewed-on: http://review.couchbase.org/69270
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>