NCBC-958: Provide sane checks for max and min pool sizes (refactor)
MOTIVATION
----------
After the initial fix that was made to add validation of the MaxSize and
minSize connection values, it was found in some situations those
values are manipualted outside of the constructor. This caused
compilation errors as the setter was made private as part of the change.

This change is to refactor the original change to allow external
modification of the max and min values, but continue to offer validation
on the values.

MODIFICATIONS
-------------
Changes the MinSize and MaxSize to use a private backing field. Upon
setting this public property, the same validation as before is executed.

Removed the constructor that was introduced as part of the original
change because it was no longer required.

Updated useage that used the new constructor to revert back to an
initialiser for setting the MaxSize and MinSize properties.

RESULT
------
The MinSize and MaxSize properties now expose a public setter and the
same validation is executed as before.

Change-Id: I5aeb3d98d4df093bdf9361a167c84260de586800
Reviewed-on: http://review.couchbase.org/69278
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>