NCBC-958: Provide sane checks for max and min pool sizes
MOTIVATION
----------
We should validate the min and max connections values before setting the
internal backing field values otherwise it is possible to catch the
exception and leave the value in an invalid state.

We also need to allow 0 minimum connections for backward compatability
with existing setups.

MODIFICATIONS
-------------
Validate the minSize and maxSize before setting the backing fields. This
is also done during the constructor where a non-defualt value is used.

Allow 0 for minimum connections.

RESULT
------
The private backing fields are not changed if an invalid min or max
connection value is used.

Minimum connections can now be 0.

Change-Id: Ibd8ce166241a03812a72b9cbf56529c5dedcdc6f
Reviewed-on: http://review.couchbase.org/69336
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>