NCBC-1079: Add streaming N1QL response support
Motivation
----------
The current N1QL query client loads the entire response payload into
memory before returning the results. For small requests this isn't a
problem, for larger responses it invariable leads to
OutOfMemoryException's (OOM) when the collection storing the results
resizes and can no longer allocate contiguous memory for the collection.
This commit adds a streaming option which reads and serializes the JSON
response per item as the server returns the bytes.

Modifications
-------------
Added a a UseStreaming(bool) method to IQueryRequest for setting the
behavior to streaming. Added another property called IsStreaming to
determine if the response should be streamed. Added a
StreamingQueryReponse class that will read each item from the current
stream. IQueryRequest now implements IDisposable so that the response
stream can be disposed by the application after reading through the
stream.

In order to use any properties on the StreamingQueryResponse, you
must first iterate the result or call the ForceRead() method which will
iterate through the results creating an internal collection with all the
items. Importantly, understand this is essentially the same behavior as
QueryResult (non-streaming) and may lead to OOM exceptions. Note that
Success and HttpStatusCode can be read without calling ForceRead or
iterating on the response. If you attempt to call any other property on
the response object without call ForceRead or iterating over it, then an
exception will be thrown.

Results
-------
When IQueryRequest.UseStreaming(bool) is set to true, the client will not
buffer the entire response, but instead keep the stream open and read each
item on demand as you iterate through the results.

Change-Id: I9683df27a0429dff49793ac3d0c3dd9d7f171851
Reviewed-on: http://review.couchbase.org/69677
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>