NCBC-1229: ScanWait does not append wait suffix
MOTIVATION
----------
When executing a N1QL query with it is possible to apply a ScanWait with
a Timespan as a max wait time. However when this is used the value that
is sent to the server is missing the 'ms' suffix to indicate the time
unit.

MODIFICATIONS
-------------
Format the scan_wait value to include the ms time unit when aa value is
provided.

Add unit test to prove the scan_wait value includes the ms suffix.

Add integration test to prove a query with scan_wait is accepted by
the server.

RESULT
------
ScanWait will send the wait time with a 'ms' suffix so the server
accepts it.

Change-Id: Ie6344390f2b4c0df81962272b623d14560234153
Reviewed-on: http://review.couchbase.org/69729
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>