NCBC-1226: BooleanQuery JSON structure is invalid
MOTIVATION
----------
It has been found that the generated JSON when building a FTS
BooleanQuery query did not conform to the expected structure by the
server. The main problem was each sub-query was always being
wrapped in a "query" object. This worked for top level queries but
not when they are nested, eg BooleanQuery which contains must, must_not
and should arrays of child-queries.

In addition to this, some fields that should be omitted from the JSON
if they were unassigned or had a default value were still being
generated and sent to the server which could lead to unknown behaviour.

MODIFICATIONS
-------------
- Remove SearchQuery.Export(ISearchParams) as it is not required.
- Update SearchClient to build the query JSON using the SearchParams,
  then add the "query" property using the Query.Export() method
- Update FtsQueryBase to default Boost to 1.0, make _boost private and
  make the Export() method create the initial JObject and apply "boost"
  if not default
- Update all Query implementations to remove Export(ISearchParams) and
  Boost overrides, use base.Export() to get the initial JObject and add
  it's properties to that
- Omit query properties that are optional where appropriate; field,
  analyzer, etc
- Update unit tests to fit with new pattern & expected output JSON
- Add unit to prove each query type's expected JSON output (TODO)

RESULT
------
The output JSON is what the server expects, particaulary around when
using nested queries such as BooleanQuery.

Fields that are optiontal are omitted from the JSON if they are
unassigned or have a default value.

Change-Id: I5692938d8ccf00c6ae35c2959c83dc8caaf428e6
Reviewed-on: http://review.couchbase.org/69486
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>