NCBC-1104: When SSL certificate is not setup correctly System.ArgumentException is thrown
MOTIVATION
----------
When the client tries to connect to a server using an SSL certificate
and the certificate is invalid, an AuthenticationException generated
and caught but only logged. This allows the connection to be used and
throw more errors when trying to send data because the connection has
not been authenticted.

MODIFICATIONS
-------------
Catch and log both AggregateException and Exception but 'throw'
afterwards to allow it bubble up and not be ignored.

Remove the _timingEnabled field as it was not being used.

RESULT
------
AuthenticationExceptions generated when trying to authenticate against a
server using an invalid SSL certificate bubble up and are not ignored.

Change-Id: I7f8ffacba1d29c6653fddd4563d448f8561f46a0
Reviewed-on: http://review.couchbase.org/70020
Tested-by: buildbot <build@couchbase.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>