NCBC-1237: Add custom "sort" param for FTS
MOTIVATION
----------
As part of the search service it will be possible to provide a list of
strings that the server will use to sort the results by before returning
to the client. The strings can be prefixed with '-' to indicate a
decending nature.

MODIFICATIONS
-------------
- Add Sort method to ISearchParams, SearchParams and SearchQuery
- Add sort fields to output JSON when executing ToJson on SearchParams
- Add unit test to prove adding sort fields are added to the JSON that
  is sent to the server

RESULT
------
It is now possible to provide sort fields to the server when executing a
search query.

Change-Id: I3bdf636ac1b614295b9f3e319a178e42576a4a7c
Reviewed-on: http://review.couchbase.org/70166
Tested-by: buildbot <build@couchbase.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>