NCBC-1232: .NET client may not automatically re-prepare when an index is removed
MOTIVATION
----------
When a N1QL query fails because the index is removed on the server that
a stored query plan is stored against, the client does not eject the
bad query plan. Originally the server responded with a generic error
code of 5000 with a custom error message. Now the discrete error
IndexNotFound (12016) is returned.

MODIFICATIONS
-------------
- Add IndexNotFound (12016) to ErrorPrepared enum
- Update the CheckRetry method on QueryClient to check for new error
  code
- Add integration test to simulate scenario. A HTTP debugger is required
  to validate because there is not easy way to mock respones within the
  client

RESULT
------
When the server responds with the error IndexNotFound (12016) any stored
query plan is ejected and a new query plan will be created.

Change-Id: I187d20c84d8366edfa375b649d1c0b39e8ff1081
Reviewed-on: http://review.couchbase.org/70249
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>