NCBC-1240: Throw InitializationException if Initialize() is not called
Motivation
----------
The ClusterHelper requires that Initialize() be called so that the current
configuration can be applied to the cached Cluster instances. If the
ClusterHelper is used without calling Initialize() a
InitializationException should be raised.

Modifications
-------------
Change ClusterHelper so that GetBucket calls will throw a
InitializationException if Initialize() has not been called first.
Refactored GetBucket overloads into a a single method to keep things DRY.

Removed comment from ClusterHelperTests incorrectly stating that the
tests depended upon localhost instance of CB; refactored one test so that
it used the TestCOnfiguration.CurrentConfiguration() which is the current
app.config configuration. Removed one test which was redundent.

Results
-------
If GetBucket is called before Initialize(), an InitializationException
will be thrown.

Change-Id: Id8efbab44a66e1d563a757903d8d586f380a320b
Reviewed-on: http://review.couchbase.org/70367
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: buildbot <build@couchbase.com>