NCBC-1234: Develop build, package, deploy scripts for .NET SDK
MOTIVATION
----------
Our release process is completely manual at the moment and because of
that is prone to human error. The process can be improved by automating
tasks such as unit testing, signing assemblies and building /
publishing packages.

MODIFICATIONS
-------------
- Add appveyor.yml
- Add build-utils\Couchbase.snk.enc (encrypted version of signing key)
- Update Couchbase.nuspec to include the $version$ flag, name the
  package and identify the output target files

RESULT
------
As new commits are pushed to GitHub from Gerrit, AppVeyor will trigger
to build the source code and execute the unit tests. When GitHub is
tagged for a release additional tasks are performed:
- Decrypt the Couchbase.snk file
- Sign the Couchbase assembly during build
- Generate nuget & .zip packages
- Create GitHub Release and attach .zip
- Publish Nuget package

NOTE: Publishing the nuget package is currently disabled because you can't
remove a nuget package once published.

Change-Id: Id1c395daf565ec21b52026fed03fb6c73496ee32
Reviewed-on: http://review.couchbase.org/70565
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>