NCBC-1245: Remove local nuget config and targets files
MOTIVATION
----------
There are local nuget config and targets files that are not needed and
they are causing trouble with the Automated CI process AppVeyor. These
should be removed as they are not required.

MODIFICATIONS
-------------
- Remove src/.nuget/nuget.targets
- Remove src/.nuget/nuget.config

RESULT
------
There is no longer a local version of either nuget.config or
nuget.targets in the repository.

Change-Id: I3a06b204d0c58b27c2a2414664816de716e2ac6b
Reviewed-on: http://review.couchbase.org/70690
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>