NCBC-864: Add JSON streaming ViewClient
MOTIVATION
----------
The default ViewClient processes all the rows of a view result into one
large collection and returns this. This is ok for small result sets,
however, computing these results in this way for larger results sets
(both number of rows and size of row) is both memory and CPU expensive.

It would be better to introduce a streaming version of the ViewClient
that can read each row one at a time, minimizing the amount of up-front
memory and CPU used.

MODIFICATIONS
-------------
- Add StreamingViewClient that supports streaming rows instead of
  processing the whole result set immediately
- Add UseStreaming method to ViewResult to instruct it to use streaming
- Add ViewClientBase to share logic between the original ViewClient and
  the new StreamingViewClient
- Add unit tests for ViewResult to check streaming is enabled
- Add Unit Tests for StreamingViewClient to check it can process results
  using streaming (includes test data)
- Fixed some BucketManager Tests that used an odd namespace that
  conflicted with View tests

RESULT
------
It is now possible to utilze streaming when executing a ViewQuery to
optimize the amount of memory and CPU used per row.

Change-Id: Ia8841232e846bd323d73a256a11c63660a44ce77
Reviewed-on: http://review.couchbase.org/71191
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>