NCBC-1254: Revert IQueryResult<T> refactoring
MOTIVATION
----------
As part of NCBC-1079 the IQueryResult<T> interface was refactored with a
new IQueryResult interface being created to hold the non-generic
properties. However, this broke integration projects that used the public
interface explicitly.

MODIFICATIONS
-------------
- Return the properties that were moved to IQueryResult to
  IQueryResult<T>
- Remove IQueryResult
- Replace direct IQueryResult usage with IQueryResult<T>

RESULT
------
The IQueryResult<T> interface is now in the expected state and
integration projects, such as Linq2Couchbase, can use this library.

Change-Id: I7703ae9435d0dfcf28ea63e5a660575a495f7e09
Reviewed-on: http://review.couchbase.org/71188
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>