NCBC-1142: Query does not recover after service restart
MOTIVATION
----------
When a N1QL query is submitted to a cluster that is restarting, the
available Query URIs can be exhausted through retries until none
are left. However, when this happens an empty URI string is returned
which causes a InvalidOperationsException when trying to send to an
invalid URI with the HttpClient. Also, no logging is available to
indicate what has happened.

MODIFICATIONS
-------------
- In the case the QueryClient doesn't retrieve a valid URI when trying
  to submit a query, log out an error and return a failure result
- Add log message to ExceptionUtils for empty N1QL URIs
- Refactor GetRandom in ArrayExtensions to remove a duplicate and make
  the main implementation more clear

RESULT
------
If the QueryClient does not retrieve a valid URI when trying to submit a
query, it logs out what has happened and returns a failed request to
prevent further retries.

Change-Id: Ifce8731190814bd2784b4822c1a5176ef7aa4a79
Reviewed-on: http://review.couchbase.org/71189
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>