NCBC-1262: Improve Query request URI handling
Motivation
----------
The client previously used a random distribution for determining which
cluster node to send a query to. This commit changes to a round-robin
approach. Additionally, there was a bug where if a Uri was not available
the client would try to use null; this has been changed so that an
InvalidOperationException will be returned instead.

Modifications
-------------
Added null check incase GetQueryUri returns null. Also added a round robin
algorithm to help distribute the load. Added ability to set custom
QueryFailedThreshold in the config (Default value is still 2).

Results
-------
N1QL queries will be sent round-robin to each query node in the cluster;
if a valid Uri cannot be obtained, then an InvalidOperationException
will be returned to the application.

Change-Id: I4b3ab3bdda0b7d6eea8eb06d6522a5098a2071c0
Reviewed-on: http://review.couchbase.org/71426
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>