NCBC-1261: Ensure (N1QL & View) streams are correctly disposed
MOTIVATION
----------
When closing both the N1QL and View streaming results the JsonTextReader
and underlying memory stream should be closed correctly to prevent
holding onto resources. Currently View streaming does not explicitly
dipose of the reader or stream and N1QL results directly disposes of the
stream but not the reader.

MODIFICATIONS
-------------
- During Dispose call JsonTextReader.Close in StreamingQueryResult
- Add IDisposeable and call JsonTextReader.Close in StreamingViewResult

RESULT
------
Both N1QL and View streaming results correctly dispose of the
JsonTextReader and underlying memory stream.

Change-Id: I60e8cbd5ee747ec8112fb9fa4904dc2410587380
Reviewed-on: http://review.couchbase.org/71434
Tested-by: buildbot <build@couchbase.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>