NCBC-1265: Update BinaryToJsonTranscoder.cs to be net45 only
MOTIVATION
----------
Some of the using statements in BinaryToJsonTranscoder.cs are only
available when compiling for .NET 4.5. The main body of the class is
already covered with a conditional compile file for NET45 and the rest
of the file should also be included.

MODIFICATIONS
-------------
- Wrap BinaryToJsonTranscoder.cs, including using  statements, in a #if
  NET45 ..  #endif

RESULT
------
All of Update BinaryToJsonTranscoder.cs is only available to when
compiling for net45.

Change-Id: I487f2d01c88704d211894f6bb581687045c5dbe8
Reviewed-on: http://review.couchbase.org/71621
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>