NCBC-1267: Disable MuxIO when UseSsl is true
Motivation
-------------
SslConnection uses SslStream which does not allow for concurrent threads
to write to the same stream. This commit makes the client use pooled
connection when useSsl is enabled. A later commit will provide an
implementation of SslConnection which supports SSL/TLS.

Modifications
-------------
- Add UseConnectionPooling property to ClientConfiguration.
- Change signature of IOServiceFactory.GetFactory to take a
  ClientConfiguration instance
- Change scope of logic in IOserviceFactory.GetFactory so that parameter
  is used when the func is invoked.

Results
-------
If UseSsl is enabled, the client will default back to pooled connections
even if UseConnectionPooling is false.

Change-Id: Ic34caf247ba42cb72b6964a3ded7c856e42953f7
Reviewed-on: http://review.couchbase.org/71879
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>