NCBC-1274: Update AppVeyor configuration
MOTIVATION
----------
After the recent work to move the NetStandard projects to xproj,
AppVeyor doesn't restore packages for non-NetStandard projects. This is
because we are not providing a target for the restore and the default
action is to only target xproj projects

We are also building SDK version 2.4.0 which means the version is out of
date.

MODIFICATIONS
-------------
- Set the version to 2.4.0.{build}
- Set the nuget restore target to couchbase-net-client.sln

RESULT
------
The updated appveyor configuration will restore packages for all
projects and sets the assembly version to 2.4.0.#.

Change-Id: Ie8eff42405fbfbab3ea8d1088aee94a1581ef7a5
Reviewed-on: http://review.couchbase.org/72106
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>