NCBC-1281: Suppress XML Comment Missing (CS1591) build warnings
MOTIVATION
----------
The Couchbase class library generates XML documentation when built in
the Release configuration. However, there are lots of classes, methods
and properties that do not have comments and generate build warnings.

As a short-term fix, we should ignore these warnings and ticket
NCBC-1282 will be used to track adding the comments.

MODIFICATIONS
-------------
- Ignore XML Comment Missing (CS1591) build warnings for Couchbase
  project when built in the Release configuration

RESULT
------
XML Comment Missing warnings are not generated when building the
Couchbase class library in the Release configuration.

Change-Id: Id6d633b5dd32822cb24f43f9334d51958e829b84
Reviewed-on: http://review.couchbase.org/72109
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>