NCBC-1274: Update nuspec title and file references
MOTIVATION
----------
The current nuspec file needs its title updating to 2.4. Also, the
file paths reference incorrect dll and xml files and uses the net45
library in the netstandard1.5 folder.

MODIFICATIONS
-------------
- Update the package title to 2.4
- Update the file paths to correctly reference both net45 and
  netstandard1.5 libraries

RESULT
------
When creating the nuget package, the package title is 2.4 and the
correct libraries are added in the right location.

Change-Id: I87c92a3686faf96bf6f72abd55e9b65013270919
Reviewed-on: http://review.couchbase.org/72112
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>