NCBC-1274: Generate XML docs during build of Couchbase.NetStandard
MOTIVATION
----------
We currently build the XML Documentation for the Couchbase library when
we build for net45, this does not happen for Couchbase.NetStandard. The
path to the xml file for netstandard1.5 also needs updating to the
generated file.

MODIFICATIONS
-------------
- Add the xmlDoc property to Couchbase.NetStandard's project.json
- Update Couchbase.nuspec to use the update file path for the xml doc

RESULT
------
When Couchbase.NetStandard is built in the Release configuration, the
xml documentation is also created and when nuget pack is created it uses
the generated xml doc from the correct location.

Change-Id: If240bc068acab79b621f20f2a5791d42bf7e3d20
Reviewed-on: http://review.couchbase.org/72122
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>