NCBC-1274: Generate correct zip folder structure
MOTIVATION
----------
The packaging of the binaries ready for upload to S3 doesn’t generate
the correct folder structure and include the files from both the net45
and netstandard1.5 bin\Releae folders.

It would also be beneficial to not only create both the binaries zip and
nuget packages only on tags. However, these would only be deployed
when the repository is tagged.

MODIFICATIONS
-------------
- Update appveyor.yml’s before_deploy script to generate the correct
  folders in the zip file with the files from the correct location
- Remove the if that checks if the build was triggered by a repo tag

RESULT
------
When the zip folder is created, it correctly creates both net45 and
netstandard1.5 f folders with their respective dll and xml doc files.

The zip and nuget files are created for each build instead of just when
the repository is tagged.

Change-Id: I1e93e8b7baeaff7be200281591de2b8da9be69dd
Reviewed-on: http://review.couchbase.org/72123
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>