NCBC-1285: Disable HTTP 1.1 pipelining to query service, if enabled
Motivation
----------
Owing to an implementation limitation in Go's HTTP 1.1 API, pipelining
cannot be used with CloseNotifier, which query uses.  This is tracked in
MB-22184. To avoid correctness issues, disable HTTP pipelining in the .NET
client, if enabled.

Modifications
-------------
Set AllowPipelining to false in the Net45 ctor of
AuthenticatingHttpClientHandler. When built for .NET Core,
AuthenticatingHttpClientHandler inherits from HttpClientHandler which
does not have an AllowPipelining property.

Results
-------
HTTP Pipelining will be disabled for Net45 builds.

Change-Id: I2127cdc1e28734ff531dfd9c434677e5e17e5f19
Reviewed-on: http://review.couchbase.org/72214
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>