NCBC-1288: NullReferenceException thrown unless the LogManager is configured.
Motivation
----------
If the LogManager is not configured, then a NRE will be thrown when the
client is instantiated.

Modifications
-------------
This commit fixes that issue by removing the
static ILog field in ClientCOnfiguration and ensuring that a default
LoggerFactory is always assigned. I also removed the ILog parameter from
TimerFactory.GetTimer and instead use a private static field to assign the
Log instance to timers. Finally, I bumped NetStandard version to 1.6.1 for
consistency.

Result
------
A NRE will not be thrown if the LogManager isn't configured.

Change-Id: I9dad8217b65fa63e0f45682b623498d52bc9f118
Reviewed-on: http://review.couchbase.org/72482
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>