NCBC-1290: Reactivate query nodes after failure
Motivation
----------
Once a query node is marked as failed, it is never reactivated and
cannot be used by the client.

Modifications
-------------
When all query nodes are marked as failed, clear the failure on all
nodes before selecting a node.

Results
-------
If all query nodes get flagged as failed, then it will retry them all to
see if any are functional again.

This is not a complete fix, as a node will not be reactivated until all
query nodes have failed or there is a config change (i.e. rebalance).

Change-Id: Ifbe34d14c3f84c7d6af5e443c3a83dcfba28b062
Reviewed-on: http://review.couchbase.org/72504
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>