NCBC-1294: Add Couchbase.snk to gitignore
MOTIVATION
----------
The Coucbase library uses an SNK to strongly name the assembly for
public use. The SNK is intended to be private to ensure that any
consuming application can guarantee that it is an official distribution.

To help prevent accidentally adding the SNK file to source control we
can add it to the project's .gitignore file.

MODIFICATIONS
-------------
- Add Couchbase.snk to .gitignore

RESULT
------
If present, the private Couchbase.snk file will not be added to source
control accidentally.

Change-Id: I55f8a51dcb14e4209b3ec6513ee270fa240b9c61
Reviewed-on: http://review.couchbase.org/72882
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>