NCBC-1295: MaxConnectionsPerServer not supported on linux
Motivation
----------
If MaxConnectionsPerServer is set for .NET Core and the host server has
libcurl < 7.30.0 installed, a PlatformNotSupportedException may be
logged. Furthermore, any values assigned to DefaultConnectionLimit will be
ignored.

Modifications
-------------
Add code comments indicating that this setting is not supported for
versions of libcurl < 7.30.0

Result
------
If logging is set for DEBUG, a PlatformNotSupportedException will be
logged. The value for DefaultConnectionLimit will be ignored.

Change-Id: I28400bd0347f5e3a4101b5eb67fb7a9a1f997763
Reviewed-on: http://review.couchbase.org/72908
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>