NCBC-1274: Fix appveyor configuration
MOVTIVATION
-----------
When commits and tags are added to GitHub the AppVeyor integration is
responsbile for building the Couchbase assemblies ready for packaging
and distribution.

MODIFICATIONS
-------------
Updated appveyor.yml to do the following:
- only build the Couchbase and Couchbase.NetStandard projects
- sign both assemblies using the decrypted Couchbase.snk file
- do not run test projects (testing is conducted on Jenkins)
- tidy up script blocks into discrete commands for easier reading
- push the zip file to Github when triggered for a tag

RESULT
------
When AppVeyor is triggered, a zip and nuget package are created with
the signed assemblies and stored as atrifacts. If AppVeyor was
triggered because of a Git Tag, a GitHub Release is created and the
zip file is also attached.

Change-Id: I4cc10b5e7d25d7e0cab1f115dd0414911eeaf499
Reviewed-on: http://review.couchbase.org/72951
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>