NCBC-1291: Reactivate search nodes after failure
Motivation
----------
Once a search node is marked as failed, it is never reactivated and
cannot be used by the client.

Modifications
-------------
When all search nodes are marked as failed, clear the failure on all
nodes before selecting a node.

Move failure threshold magic number to a constant.

Results
-------
If all search nodes get flagged as failed, then it will retry them all to
see if any are functional again.

This is not a complete fix, as a node will not be reactivated until all
search nodes have failed or there is a config change (i.e. rebalance).

Change-Id: I73974988c923a8999d691ec2d4e49cfbfdeafb85
Reviewed-on: http://review.couchbase.org/73277
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>