NCBC-1249: Add Analytics Query Support as Uncommitted
MOTIVATION
----------
An upcoming server feature will be Couchbase Analytics (CBAS). This is a
new service with an API very similar to the N1QL API. This is the first
commit to introduce the new Analytics bucket level queries.

MODIFICATIONS
-------------
- Add AnalyticsRequest, AnalyticsResult with supporting interfaces and
  data mapping classes
- Add Query and QueryAsync methods to Bucket, and supporting
  implementations
- Extend Configuration to support retrieving an Analytics service URIs
- Add basic tests to illustrate usage

RESULT
------
It is possible to submit an analytics query via an IBucket. Only simple
queries are possible right now.

NOTES:
Analytics requests currently do not support:
- Positional or named parameters
- Prepared statements and/or caching
- Cluster level querying
- The query endpoint is always enabled, and uses hardcoded ports 8095
  or 18095 for SSL

Change-Id: I934b75d5edb77b4bc490eb7482d78ac6d0940453
Reviewed-on: http://review.couchbase.org/73315
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>