NCBC-1311: Update GetWithLock and GetAndLock expiration comments
MOTIVATION
----------
When a GetAndLock operation is sent to the server with an expiration
value of 0, the server default lock expiration is used (15 seconds).
However the code comments for GetAndLock and GetWithLock both indicate
an infinite timeout would be used on the server.

MODIFICATIONS
-------------
- Update GetAndLock and GetWithLock code comments to indicate the server
  default lock expiration is used if an expiration of 0 is given

RESULT
------
The code documentation accurately reflects what lock expiration value
will be used by the server.

Change-Id: Ic3caf3bc5cb7b3dd8e1a914bf4314c0ce975c1e6
Reviewed-on: http://review.couchbase.org/73579
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>