NCBC-1315: Send key with hello operation
MOTIVATION
----------
When the Hello operation is sent to the server, it expects the key
section of the packet to be present because it uses it to log which SDK
has sent the request. This is useful for debugging purposes.

Currently the key is not sent and so the server does not log which SDK
has sent the request.

MODIFICATIONS
-------------
- Move custom hello body logic into CreateBody() override
- Remove Write() override
- Add unit test to ensure key is written during the Write() method call

RESULT
------
The key is sent in the hello operation packet.

NOTE: There are existing integration tests to prove that the Hello
operation works as expected.

Change-Id: Ic3a8881e5fcb8574f134755319b0a42c05788cd5
Reviewed-on: http://review.couchbase.org/73698
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>