NCBC-1322: Test that DocumentFragment<T>.Value works as expected for POCO's
Motivation
----------
DocumentFragement<T>>Value is flagged as Obsolete: "For backwards
compatibility with regular K/V and Document API internals. Do not use.".
However, it is still possible to call it and retrieve a list of
ICollection<OperationSpec> if cast to DocumentFragment<T> where T is the
type of the DocumentFragment<T> generic parameter. Otherwise, null will be
returned if LookupIn.Get(..).Execute() is called without casting. The
purpose of the tests is to document this behavior since it is
un-intuitive.

Modifications
-------------
Added tests to Couchbase_SubDocument_Tests.cs illustratating behavior.

Results
-------
Do not use DocumentFragment<T>.Value as it has been flagged as obsolete
and may be removed in a future major release.

Change-Id: I15c81bbf5b99b74c90796a69e83e10d651d8d032
Reviewed-on: http://review.couchbase.org/73762
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>