NCBC-1324: Remove InternalsVisibleTo before build in AppVeyor
MOTIVATION
----------
When AppVeyor builds the Couchbase project it tries to also sign the
output assemblies. When doing that with InteralsVisibleTo entries, each
'friend' project must also include a public key. It's not necessary to
include the friend assemblies as they are only used during development
and testing so can be safely removed before building.

MODIFICATIONS
-------------
- Update appveyor.yml to remove any lines that contain
  InternalsVisibleTo in AssemblyInfo.cs before building

RESULT
------
Before AppVeyor builds the Couchbase Project, it removes the
InternalsVisibleTo entries in the AssemblyInfo.cs.

Change-Id: I5857c5327ced3efc34cd833a2981db541111e840
Reviewed-on: http://review.couchbase.org/73821
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>