NCBC-1325: Move packing steps to after_build in AppVeyor config
MOTIVTION
---------
We want to create a zip and nuget package on each commit, with them
being deployed on a git tag. Currently the steps to do the packing work
is done in before_deploy but this looks to only get executed if there is
a deployment task.

MODIFICATIONS
- Moved the packaging commands in before_deploy to after_build in
  appveyor.yml

RESULT
------
A zip and nuget package should be created for each commit as it runs
through AppVeyor.

Change-Id: I7d116e25d1111ab03fb75e5f01ed6e25f3afcc67
Reviewed-on: http://review.couchbase.org/73839
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>