NCBC-1305: Make bootstrap url default to port 8091
Motivation
----------
If no port is provided, the System.Uri class will default to port 80. To
bootstrap, Couchbase expects port 8091, so we we need to ensure that the
default is port 8091 instead of port 80.

Modifications
-------------
Check for port 80 when ClientConfiguration.Initialize is called and
rewrite the uri to use port 8091.

Results
-------
If no port is provided in the Uri, the sdk will default it to port 8091
instead of port 80.

Change-Id: Iaec736d79388a76ca6f5f7f8cfb8de2d766e1d4c
Reviewed-on: http://review.couchbase.org/73869
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>